requirements.txt
faiss-cpu==1.7.4
numpy==1.26.4
sentence-transformers==3.0.1
huggingface-hub==0.24.5
openpyxl==3.1.5
streamlit==1.49.1
pandas==2.2.2
langchain-core==0.3.76
langgraph==0.6.7
google-ai-generativelanguage==0.6.18
langchain-google-genai==2.1.10
